requirements_workflow.txt
# Event-Driven RAG Workflow - Python Dependencies
# Install with: pip install -r requirements_workflow.txt

# Core LlamaIndex & Workflows
llama-index-core>=0.10.0
llama-index-workflows>=0.1.0

# Vector Store & Embeddings
pinecone-client>=3.0.0
llama-index-vector-stores-pinecone>=0.1.0
llama-index-embeddings-cohere>=0.1.0

# LLM Providers
llama-index-llms-openai>=0.1.0

# Data & Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async Support
aiohttp>=3.8.0
asyncio-contextmanager>=1.0.0
